+++ src/pages/SYha/Main/components/Aside_left/Feed.scss
@@ -1,0 +1,137 @@
+@import '../../../styles/mixin.scss';
+@import '../../../styles/variables.scss';
+@import '../../../styles/common.scss';
+
+.feed {
+  height: 100%;
+  border: 1.5px solid $lineColor;
+  margin-bottom: 50px;
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  background-color: white;
+
+  .feedhead {
+    height: 8%;
+    padding: 10px;
+    position: relative;
+    font-size: 17px;
+    font-weight: 900;
+
+    .profile_img {
+      border-radius: 100%;
+      width: 30px;
+      height: 30px;
+      margin-right: 10px;
+    }
+    .dropdownmenu_iconimg {
+      width: 20px;
+      height: 20px;
+      margin-left: 300px;
+      position: absolute;
+      right: 10px;
+    }
+  }
+
+  .feedimg {
+    height: 63%;
+    margin: 0px;
+    object-fit: cover;
+  }
+
+  .reaction_icon_wrap {
+    height: 5%;
+    padding: 0px 10px;
+    position: relative;
+
+    .reaction_iconimg {
+      width: 18px;
+      height: 18px;
+      margin: 5px 8px;
+    }
+
+    .saveicon {
+      position: absolute;
+      right: 10px;
+    }
+  }
+
+  .show_like_wrap {
+    height: 3%;
+    padding: 0px 5px;
+    font-size: 12.5px;
+
+    strong {
+      font-weight: 700;
+    }
+
+    .friendProfile_img {
+      width: 20px;
+      height: 20px;
+      margin: 0px 8px;
+      border-radius: 100px;
+    }
+  }
+
+  .feed_discription_wrap {
+    margin: 2px 10px;
+    font-size: 12.5px;
+
+    .discription {
+      margin: 5px 3px;
+      position: relative;
+
+      strong {
+        font-weight: 700;
+      }
+    }
+
+    // .timecount {
+    //   color: $lineColor;
+    //   margin: 0px 4px;
+    // }
+  }
+  .comment_wrap {
+    overflow: auto;
+    max-height: 60px;
+  }
+  .addcomment {
+    font-size: 15px;
+    width: 100%;
+    position: relative;
+    position: absolute;
+    bottom: 0px;
+
+    .commentinputbox {
+      width: 100%;
+      border-style: none;
+      height: 45px;
+      border-top: 1px solid $lineColor;
+      padding: 10px;
+    }
+
+    .button:disabled {
+      color: #c4e1fb;
+    }
+    .button {
+      background: white;
+      color: #64acf3;
+      border: none;
+      padding: 0;
+      font: inherit;
+      font-weight: 600;
+      cursor: pointer;
+      outline: inherit;
+      position: absolute;
+      right: 20px;
+      top: 40%;
+    }
+  }
+}
+
+// .joayo {
+//   width: 18px;
+//   height: 18px;
+//   position: absolute;
+//   right: 5px;
+// }
